+++ docs/conf.py
@@ -8,7 +8,7 @@ copyright = "2024"
 author = "esophagoose"
 
 release = "0.1"
-version = "0.10.0"
+version = "0.10.1"
 
 extensions = [
     "sphinx.ext.duration",